+++ src/app/loadout-drawer/loadout-drawer-reducer.test.ts
@@ -21,7 +21,7 @@ import {
   toggleEquipped,
   updateMods,
 } from './loadout-drawer-reducer';
-import { filterLoadoutToAllowedItems, newLoadout } from './loadout-utils';
+import { newLoadout } from './loadout-utils';
 
 let defs: D2ManifestDefinitions;
 let store: DimStore;
@@ -217,16 +217,17 @@ describe('addItem', () => {
     expect(loadout.items).toEqual([]);
   });
 
-  it('removes class-specific items when saving as "any class"', () => {
-    const hunterItem = allItems.find((i) => i.classType === DestinyClass.Hunter)!;
-    expect(hunterItem).toBeDefined();
+  // TODO: reenable after https://github.com/Bungie-net/api/issues/1937 is fixed
+  // it('removes class-specific items when saving as "any class"', () => {
+  //   const hunterItem = allItems.find((i) => i.classType === DestinyClass.Hunter)!;
+  //   expect(hunterItem).toBeDefined();
 
-    let loadout = addItem(defs, hunterItem)(emptyLoadout);
-    loadout = setClassType(DestinyClass.Unknown)(loadout);
-    loadout = filterLoadoutToAllowedItems(defs, loadout);
+  //   let loadout = addItem(defs, hunterItem)(emptyLoadout);
+  //   loadout = setClassType(DestinyClass.Unknown)(loadout);
+  //   loadout = filterLoadoutToAllowedItems(defs, loadout);
 
-    expect(loadout.items).toEqual([]);
-  });
+  //   expect(loadout.items).toEqual([]);
+  // });
 
   it('does nothing if the bucket is already at capacity', () => {
     const weapons = items.filter((i) => i.bucket.hash === BucketHashes.KineticWeapons);